libta: add basic implementation for a timed automaton
Implement a class representing a timed automaton. As usual, a TA
consists of a set of states, a set of named clocks (stored in a map),
and a set of transitions (stored as multimap state -> transition).

Each transition consists of a source state, target state, a symbol, a
set of clock constraints, and a set of clock resets (stored as multimap
clock_name -> constraint). A clock constraint uses the usual operators
<,<=,==,>=,> and uses the class of std::less operators to represent
them.

The TimedAutomaton class also has a method accepts_word, which checks
whether the automaton accepts the given input word.

This is a first take on the implementation and may see major rework.

Clock resets are not implemented yet.

Also provide some basic tests to check that the TimedAutomaton works as
expected.