libta: add nested Class Path to store a run on the TA
Instead of modifying the TA itself, store run information in a separate
Path class. This avoids having a TA in a different state after calling
make_transition or accepts_word.

Also simplify the tests, as we do not need to create a new copy of the
TA every time we check if a word is accepted.