libta: compute minimal models of ATA formulas
To compute a run of an ATA (or to determine whether an ATA accepts a
word), we need to compute the minimal models of ATA formulas, as these
determine the next configuration in each step. Add the method
`get_minimal_models` to ATA formulas that determines a set of minimal
models for each formula.

This uses ranges to compute the cartesian product of two sets, which is
necessary to compute the minimal model of a conjunctive formula.