libta: add a first implementation for Alternating Timed Automata
Define Alternating Timed Automata following Ouaknine & Worrel, 2005.
An ATA uses alternating symbol and time transitions, where a symbol
transition reads a symbol and a time time transition progresses the
time. A run on an ATA is defined by a sequence of configurations, where
each configuration consists of a set of so-called states (location-time
pairs). Symbol transitions are defined as triple (source, symbol,
formula), where the source is the current location, the symbol is the
symbol to be read, and the minimal models of the formula define the
resulting configurations. Each transition may have multiple resulting
configurations, as we may have disjunctions in the formula, resulting in
multiple minimal models. Thus, the transition functions always return
vectors of runs.

This defines the low-level machinery for ATAs. The check whether an ATA
accepts a given timed word is not yet implemented.