libta: add word acceptance for ATAs
Implement the ATA method `accepts_word` based on the already existing
transitions. This directly follows the definition of a run on a word by
Ouaknine & Worrell, 2005. The function iteratively computes the
resulting runs by making transitions according to the given word and
checks whether there is a resulting configuration that only consists of
accepting locations.