ta: use a virtual function for printing, implement remaining types
We need to use a virtual function to print a formula, as a conjunction
takes pointers to two abstract formulas, for which we need to use the
sub-class's method for printing. As this is only possible using a
virtual function, declare one virtual friend operator<< for the base
class which just calls `print_to_ostream`, and implement that function for
all sub-types.