sync product: fix region increment for maxed regions
If we have a partition where some configuration has reached the max
region index but some other configuration has not yet reached the max
region index, then we need to split this partition, as we only increment
the region index of the nonmaxed subset.

This fixes a bug where a word would contain both even and odd region
indexes after an increment, because the odd max region index would not
be incremented.