automata: move function definitions into a separate hpp file
Template instantiation with `extern template class` only works if the
functions are not defined inline. As we need to separate the definition
from the class definition anyway, move the function definitions into a
separate file, to have a concise header file for the interface.

For clock constraints, we even know all the times that the template can
be instantiated with. Instantiate all types and do not include the hpp
file to forbid instantiating other types.