Move all operator<< to the global namespace and declare before use
clang is more strict that gcc and cannot find the operator overload for
types that are just templated aliases for types from the std namespace,
which we use often. Instead, the operator needs to be defined in the
global namespace. Also, the operator always needs to be declared before
being used, so we need to change the order of declarations and sometimes
even the order of includes to make it working.