ata: prune models of a disjunct if the other disjunct contains a subset
When computing the minimal models of a disjunction, we cannot simply
merge the sets of minimal models of the two disjuncts. One disjunct may
have a minimal model that is a superset of a minimal model of the other
disjunct. In that case, we need to prune the superset model and only
keep the smaller one.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>