search: when labeling a non-leaf node, consider time to reach a child
Even if it is possible to reach a bad child node with an environment
action, we may still be in a good state if we can take a controller
action before the environment action may occur. In order to detect this
case, we need to pass up the region indexes when computing the
successors.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>