Make the TA location a NamedType
The location is now a strong type and must be constructed explicitly.
This is helpful because it makes interfaces clearer, avoids errors by
passing the arguments in the wrong order (e.g., swap symbol and
location), and also allows us to define the ostream operators only for
the named types.