search: do not protect complete node with a mutex, use atomics instead
Remove the mutex that protected the complete node, as this always leads
to deadlocks. Instead, make the node's label atomic such that we can
access it from multiple threads. Also add a bool is_expanded that
indicates that a node has been expanded. This is also used to protect
the access to the node's children. This relies on the fact that we do
not modify the vector of children after the node has been expanded, as
otherwise we might get memory violations.