ta product: extend product to more than two input automata
We may want to compute the product of more than two automata. While
theoretically, this is possibly by computing (ta1 x (ta2 x ta3)), this
is not feasible in practice, as the result is a tuple of tuples.
Instead, take a vector of automata when computing the product and
directly compute the product of all the input automata.

This restricts the input automata to be of the same type. However, in
our applications, this restriction is acceptable.