search: return empty canonical word if no ATA successor exists
For a given time successor and action symbol, if there is a TA successor
but no ATA successor, we know that after executing that action, the
specification can no longer be satisfied, i.e., we have reached a good
node. Previously, this successor node would not be generated at all,
leading to wrong node labels. Now, a node with an empty canonical word
is generated, which is DEAD and thus will be labeled with TOP.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>