controller: create a controller from the labeled search tree
Each location of the controller corresponds to a location in the search
tree. The controller selects controller actions that lead to a sub-tree
labeled with TOP. Note that the controller does not contain all possible
paths in the search tree and it may contain clock constraints that are
unnecessary. However, the resulting controller is always correct.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>