sync product: test partitioning with approximate float comparison
Add a test case for the bug that we have observed: If two clock
valuations have different integer parts but the same fractional parts,
they must end up in the same partition. This used to be buggy, as we
directly compared the floats, which had slightly different fractional
parts due to different precision (depending on the integer part).