search: fix monotonic domination check in search graph
We may hit a loop, which is why we cannot just blindly move up the
graph. Instead, create a vector of seen nodes and add the current node
while moving up to the ancestors. If the current node has been seen
before, abort, as we will not find a monotonic domination in that path.