controller: skip further nodes if we have found a good controller action
When creating the controller, we can safely skip all further successors
once we have found a controller action. This is because at that point,
the controller can always follow a good path, independent of the other
sub-trees.

This keeps the resulting controller much smaller, although it now
contains less options.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>