sync product: always only have a single partition with maxed regions
Instead of keeping them separate, merge all regionalized configurations
with maxed regions into a single partition. This effectively ignores
the fractional part of all clocks that are in the max region and avoids
issues if we can reach the same configuration but with different
partitioning of the maxed clocks.

Also simplify the code, as we can now assume that there is always only a
single partition with maxed regions.