golog: wrap program in a GologProgram class
The class manages a Golog++ program and provides additional types that
we need for search, e.g., the nested Location type, and additional
methods, e.g., get_initial_configuration().

Also get rid of the test fixture for Golog, which was also mainly used
to manage the program, which is now done by the GologProgram.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>