search: implement get_next_canonical_words as part templ specialization
To get rid of the dependency on the particular Plant model, define a
generic template get_next_canonical_words, which serves as adapter
definition for all plant type instantiations. Implement the TA adapter
as partial template specialization of that template. This allows to only
include the general adapter in the search, while the TA adapter is only
needed if we actually search over a TA.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>