golog: add special terminate actions to avoid stalling
This avoids two stalling scenarios:
If either the enviroment or the controller is the only player that can
do an action, it could decide to do nothing and therefore stall the
execution.

To avoid this, add two actions env_terminate and ctl_terminate that
allow the env and the controller to terminate (respectively), therefore
avoiding the stall by the other player. Only allow these actions if the
region increment is the max region increment (i.e., it is the last
action considered).

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>