golog: directly insert terminate actions while computing successors
Instead of defining the action as part of the program, directly insert
the actions when computing the successors. This makes handling the
terminate actions much simpler, because we do not need to look up the
action in the environment, and we do not need to mess with the program
pointer (which points to a different action, not the terminate actions),
which turned out to be problematic when checking whether the program was
final.

As part of this change, we need to pass the controller and environment
actions to the successor generator functions `get_next_canonical_words`.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>