search: optionally allow set semantics for constraints
If we use location-based constraints, then we may be interested in set
semantics for MTL, i.e., the ATA reads a set of symbols (properties that
are satisfied in the current location), rather than single symbols.

To do so, add a new template parameter use_set_semantics. Also add a
dependent typedef for ATAInputType, as this is a set of
ConsstraintSymbolType's if we are using set semantics, and simply
ConstraintSymbolType if not. The type is automatically deduced using
type traits.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>