gocos: compare Golog locations based on the satisfied fluents
The history is not suitable for comparison, as the history always grows,
even if we eventually reach the same remaining program with the same
satisfied fluents (thus reaching a fixpoint). Instead, compute the
satisfied fluents and use those for comparison. To do so, add a pointer
to the program to the location, as we need to ask the program for the
satisfied fluents.