search: do not merge successors with different actions based on reg_a
This fixes a bug in the following situation: If two actions generate the
same plant successor, but with different ATA successors, then the two
successors were merged into a single successor node. However, this is
wrong, as we now can no longer distinguish which action has occurred.
Instead, we should only merge canonical words with the same reg_a if
they can be reached with the same (timed) action. If we reach a
canonical word with the same reg_a but a different action, then this
must go into a separate node.

Note that if the two successors are identical (i.e., same plant and ATA
configuration), then they will be merged regardless, as we re-use a node
if it is completely the same. This is however different to the scenario
described above, where we only compared based on reg_a. Here, we also
include the ATA configuration for comparison.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>