search: first add new children to the queue, then re-check for mon.dom.
If we have a node that has a loop to itself and is therefore
monotonically dominating, we should still add its children to the queue,
as the children may have other parents. If we do not add the children to
the queue, they will never be expanded and therefore not labeled
properly, which may result in the wrong label for the other parent.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>