search: exclude self loops from monotonic domination check
Treating self loops as monotonically dominating is dangerous: The
initial assumption of monotonic domination is that if we could reach a
bad configuration from the current node, then we can also reach that
configuration from the dominated ancestor, thus we can stop expanding
the current node. However, if the ancestor is actually the same node,
then we must not stop expansion, as this means that we check neither
node. Therefore, we must exclude self loops in the monotonic domination
check.