search: avoid infinite recursion in label function
If we have some loop of unlabeled nodes, then the label function would
recurse into the children infinitely often. To avoid this, track which
nodes we have already seen and abort the labeling if we have already
visited the current node. This means that if we encounter a loop, then
the whole loop may not obtain a label, as it is possible to just stay in
the loop indefinitely. As no node is labeled with BAD, any node that is
a parent of one of the loop's nodes may be labeled with GOOD (unless
there is some other reason why it is bad). Note that this may also mean
that the controller can just stay in the loop to obtain a winning
strategy.