benchmarks: use std::string_view in railroad benchmark
This improves the performance significantly, as we will not allocate any
strings during search.

For Railroad, this is slightly more involved, as we need to make sure
that the strings' lifetimes exceed the views' lifetimes. Therefore,
create a new class RailroadProblem that stores all the strings and that
wraps the problem specification.