mtl ata trans: correctly translate disjunctions with >2 subformulas
We made the incorrect assumption that an MTL formula always has exactly
two disjuncts. However, an MTL disjunction may contain any number of
subformulas (the subformulas are stored in a vector), which resulted in
incorrect translations. In particular, we would only consider the first
and last disjunct and ignore the rest.

Fix this by translating the MTLFormula by recursively creating
disjunctions until we have reached a single subformula. This means that
the ATA disjunction still only consists of exactly two subformulas, but
these are then again disjunctions if we are translating a disjunction
with more than two subformulas.

We may in the future decide to also have an arbitrary number of
subformulas in ATA disjunctions.

Also add a test case that failed before the bugfix.