mtl ata trans: correctly translate conjunctions with >2 subformulas
We made the incorrect assumption that an MTL formula always has exactly
two conjuncts. However, an MTL conjunction may contain any number of
subformulas (the subformulas are stored in a vector), which resulted in
incorrect translations. In particular, we would only consider the first
and last conjunct and ignore the rest.

Fix this by translating the MTLFormula by recursively creating
conjunctions until we have reached a single subformula. This means that
the ATA conjunction still only consists of exactly two subformulas, but
these are then again conjunctions if we are translating a conjunction
with more than two subformulas.

We may in the future decide to also have an arbitrary number of
subformulas in ATA conjunctions.

Also add a test case that failed before the bugfix.