gocos: introduce a clock for each action duration
Whenever a start action occurs, insert a clock and reset it to 0.
Whenever an end action occurs, compare the clock value against the
action duration. More precisely, pass the current clock value to gologpp
so it only generates a successor if the duration constraint is
satisfied.

This allows to model some kind of timing constraint in golog without
introducing full clock constraints.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>