search: adapt labeling so the controller cannot decide to do nothing
It has become clear that the controller should *not* have the choice to
do nothing. From a theoretical point of view, the controller must choose
a non-empty set of actions (which may be controller or environment
actions), the environment chooses then one of those actions. This means
that if all actions are bad, then the node is bad, because the
controller does not have any non-empty subset that it can choose that
force a good successor. On the other hand, if all environment actions
are good, then the controller may just choose all environment actions
(which satisfies the criteria that the environment may not be
restricted), thus the node should be labeled good.

This also affects two test cases that actually assumed that the
controller can decide to do nothing. In a third test case, the label
remains the same, but the label reason changes.

Co-Authored-By: Stefan Schupp <stefan.schupp@cs.rwth-aachen.de>