gocos: get rid of unnecessary terminate actions
We already enforce liveness because we no longer allow the controller or
environment to decide to do nothing. Thus, the terminate actions are no
longer necessary, the resulting labeling is already correct.

This also simplifies the test cases, because there no successors for the
terminate actions anymore. However, the label of the root node stays the
same, showing that the terminate actions are indeed no longer necessary.