gocos: use all fluents for program states, not just relevant ones
For computing the ATA successor, we filter the satisfied fluents by the
atoms that are mentioned in the MTL formula. While this is correct for
the ATA, we need to store all satisfied fluents in the GologLocation,
not just the ones mentioned in the constraint. Otherwise, two different
states of the world may compare to be equal just because they only
differ in fluents that are not mentioned in the constraint.