Check that cannonical word does not contain region > max region
If the canonical word contains a region index that is larger than the
maximal region index, then we certainly ran into some bug. As this is
difficult to spot, check this in the validity check of the canonical
word.

This may happen if the max constant passed to the search is actually
incorrect and the program or specification contains a larger constant.