sync product: reimplement time successor computation
This fixes two issues. First, if a canonical word contains an even
region index, then that region index must always be incremented first
because its fractional part is 0.

Second, when computing the time successor for a node, we need to
consider all canonical words in the node. While this was partially fixed
in a previous commit, the fix was incomplete. As an example, consider a
node with two words:
1. { (l0, c, 1) }
2. { (ATA, 0) }, { (l0, c, 1) }

Then the first time successor of (2) increments the clock of the ATA but
not of the TA. However, in word (1), there is no ATA component to
increment. Therefore, the time successor with increment 1 must be the
same word. To compute these time successors correctly, we first compute
the time successors of each word. We then check if there is any word
where the reg_a does not change. If that is the case, then we need to
keep the original word for any word where the time successor has a
change reg_a.

This makes some assumptions:
* The node only contains words with the same reg_a.
* If two words contain ATA components for different ATA locations with
  the same region index, then the clock values are assumed to be the
  same, i.e., the fractional parts are equal.