app: fix test so it can be run from anywhere
Using a path relative to `std::filesystem::current_path()` has the
disadvantage that we can run the test only from the test directory.
Passing cmake's `CMAKE_CURRENT_BINARY_DIR` to the compiler allows us to
use an absolute path, which in turn allows us to run the test from
anywhere.