golog-app: add first implementation of standalone golog tool
This implementation is derived from the implementation of the standalone
TA tool and the GOLOG Robot and GOLOG Benchmark code. It exposes the
barebone functionality for controller synthesis for GOLOG programs of
TACoS in the form of a standalone binary in the style of the original
TA tool, but requires more information (e.g. environment actions and K
number) to be supplied by the user. These functionalities will be added
later.