golog-verification: add files for path search implementation
The idea of verification for Golog programs is that if the tree search
for a given program + MTL specification sans controller actions (i.e.
every action is an environment action), leads to the root node being
labelled as top, then no execution of the Golog program violates the
given specification, i.e. the program is safe against the specification.
On the other hand if the resulting tree labels the root node as bottom,
then there is at least one unsafe execution path of the program. In this
case, finding the path gives a counter-example against the MTL spec. A
simple path traversal on a fully built search tree towards a top or
bottom labelled leaf node is to be implemented in these files, to
enable gocos to do Golog program verification.